Fix a few parameter types in cISC4PollutionSimulator

+++ gzcom-dll/include/cISC4PollutionSimulator.h
@@ -17,22 +17,22 @@ class cISC4PollutionSimulator : public cIGZUnknown
 
 		virtual cISC4SimGrid<short>* GetPollutionGrid(uint32_t dwPollutionType) = 0;
 		virtual uint16_t GetPollutionValue(uint32_t dwPollutionType, uint32_t dwCellX, uint32_t dwCellZ) = 0;
-		virtual bool GetAirValue(uint32_t dwCellX, uint32_t dwCellZ, int16_t& wValue) = 0;
-		virtual bool GetWaterValue(uint32_t dwCellX, uint32_t dwCellZ, int16_t& wValue) = 0;
-		virtual bool GetGarbageValue(uint32_t dwCellX, uint32_t dwCellZ, int16_t& wValue) = 0;
+		virtual bool GetAirValue(uint32_t dwCellX, uint32_t dwCellZ, int32_t& wValue) = 0;
+		virtual bool GetWaterValue(uint32_t dwCellX, uint32_t dwCellZ, int32_t& wValue) = 0;
+		virtual bool GetGarbageValue(uint32_t dwCellX, uint32_t dwCellZ, int32_t& wValue) = 0;
 
 		virtual int32_t GetAverageAirValue(void) = 0;
 		virtual int32_t GetAverageWaterValue(void) = 0;
 		virtual int32_t GetAverageGarbageValue(void) = 0;
-		
+
 		virtual bool IsAirPolluted(uint32_t dwCellX, uint32_t dwCellZ) = 0;
 		virtual bool IsWaterPolluted(uint32_t dwCellX, uint32_t dwCellZ) = 0;
 		virtual bool IsGarbageAccumulated(uint32_t dwCellX, uint32_t dwCellZ) = 0;
 		virtual bool IsRadioactive(uint32_t dwCellX, uint32_t dwCellZ) = 0;
 
-		virtual bool GetWorstAirTract(uint32_t& dwCellX, uint32_t& dwCellZ, int16_t& wValue) = 0;
-		virtual bool GetWorstWaterTract(uint32_t& dwCellX, uint32_t& dwCellZ, int16_t& wValue) = 0;
-		virtual bool GetWorstGarbageTract(uint32_t& dwCellX, uint32_t& dwCellZ, int16_t& wValue) = 0;
+		virtual bool GetWorstAirTract(uint32_t& dwCellX, uint32_t& dwCellZ, int32_t& wValue) = 0;
+		virtual bool GetWorstWaterTract(uint32_t& dwCellX, uint32_t& dwCellZ, int32_t& wValue) = 0;
+		virtual bool GetWorstGarbageTract(uint32_t& dwCellX, uint32_t& dwCellZ, int32_t& wValue) = 0;
 
 		virtual uint32_t GetGarbageSupply(void) = 0;
 		virtual int32_t GetLandfillTileCapacity(void) = 0;
@@ -66,15 +66,15 @@ class cISC4PollutionSimulator : public cIGZUnknown
 		virtual uint32_t GetWasteToEnergyBuildingCount(void) = 0;
 		virtual uint32_t GetRecyclingCenterCount(void) = 0;
 		virtual uint32_t GetWaterTreatmentPlantCount(void) = 0;
-		
+
 		virtual bool GetGarbageBuildings(SC4List<cISC4Occupant*>& sList) = 0;
 		virtual uint32_t GetBuildingGarbageCapacity(cISC4Occupant* pOccupant) = 0;
 		virtual float GetGarbageScalingFactor(void) = 0;
-		
+
 		virtual bool IsActiveLandfill(uint32_t dwCellX, uint32_t dwCellZ) = 0;
 		virtual bool SetIsActiveLandfill(uint32_t dwCellX, uint32_t dwCellZ, bool bActive) = 0;
 		virtual int64_t GarbagePutInLandfill(int32_t nGarbage) = 0;
-		
+
 		virtual bool GetMaxCapacity(uint32_t dwProperty, cISC4Occupant* pOccupant, uint32_t& dwResult) = 0;
 		virtual bool GetCurrentCapacity(uint32_t dwProperty, cISC4Occupant* pOccupant, uint32_t& dwResult) = 0;
 
@@ -83,7 +83,7 @@ class cISC4PollutionSimulator : public cIGZUnknown
 
 		virtual bool GetCurrentOccupantCondition(cISC4Occupant* pOccupant, SC4Percentage& sCapacity) = 0;
 		virtual bool GetCurrentOccupantEfficiency(cISC4Occupant* pOccupant, SC4Percentage& sCapacity) = 0;
-		
+
 		virtual bool GetIncineratorUsedCapacity(cISC4Occupant* pOccupant, uint32_t& dwCapacity) = 0;
 		virtual bool GetWasteToEnergyUsedCapacity(cISC4Occupant* pOccupant, uint32_t& dwCapacity) = 0;
 
@@ -95,7 +95,7 @@ class cISC4PollutionSimulator : public cIGZUnknown
 		virtual bool GetWaterTreatmentUsedCapacity(cISC4Occupant* pOccupant, uint32_t& dwCapacity) = 0;
 
 		virtual bool CalcOccupantMaintenanceCost(cISC4Occupant* pOccupant, uint64_t& dwCost) = 0;
-		
+
 		virtual void AddRadiationToCity(cS3DVector3& sCenter, float fRadiusX, float fRadiusZ) = 0;
 
 		// Does this belong here? It's in the vtable...
